Fix typo + handle lack of permission better
Before lack of "Manage Webhooks" permission error was handled like every other - it just dropped you at the server list. Now, you get a separate message if the bot doesn't have the "View Channel" or the "Manage Webhooks" permissions.
Also changed "webhookID" to "webhookId" to fix a typo and a bug that happened because of it